Fix the bug regarding exceeding column 72 with UTF-8 Japanese literal (#621)
Fixed the bug regarding exceeding column 72 with UTF-8 Japanese literal.

detail:

In the case of UTF-8 Japanese strings are used, '\n' would be inserted to column 73.
Due to the difference in the number of bytes between SJIS and UTF-8, '\n' may be inserted in a different position than in SJIS, resulting in an error.